broke down sds_data_manager_stack into components
